trento: Remove Trento docs from maintenance/15_SP1 branch
Trento, as a containerized app, is independent from the SLE release
cycle. It does not make sense to publish its docs individually for each
SLES-SAP version and neither does it make sense to backport to older
branches.

I am leaving the solution of using main (Trento docs development) and
maintenance/15_SP3 (Trento stable docs->dsc) intact for the moment --
it's not 100% logical but it's good enough and one level of backports
is not that much extra effort.